Vagrant 'development playground' environment
This adds a multi-node Vagrant environment consisting of a tron 'master' node
and three 'batch-*' nodes for jobs/services.

The master node is fully configured, and ready for addition of the nodes and node_pool
entries via tronfig, which is documented in docs/developing.rst.

The tron master node is currenly Ubuntu 10.04 LTS, due to limitations in the debian package
build process. This also mirrors the most common deployment pattern for tron at the time of
writing, hopefully with a view to making this easier to migrate from.