Add per-job/service owner/summary/notes fields
With a large tronfig managed by many people, it can be difficult
to know who should be contacted when performing maintenance.

Namespaces help with this, but it is still hard without going to the job
code itself, or git blame.

This allows job/service authors to put contact details, a summary of the job/services
purpose, and any support notes they feel are relevant, by adding the following optional fields
to the job and service schemas:

* 'owner' (string)
* 'summary' (string)
* 'notes' (string)