reset actionrun endtime on recovery
each time an actionrun enters a terminal state (as it does with unknown)
an end time is set for the job.

this is particularly confusing when a job is set back into a running
state by the recovery job, since the job will be marked as 'running' but
also have an end time. this change resets the end time to None, such
that the job appears 'in progress' again.