handle failures streaming  to stdout/stderr
if stdout or stderr disappear from under us and the thread reading them
crashes, then the stderr and stdout of the subprocess being run can
backup, and possibly prevent it exiting (see TRON-339 for an example of
docker doing this). catch any exceptions we get writing so that we
continue to read from the stderr and stdout of the subprocess.