Add Service Account support for k8s jobs
We'll use this for Pod Identity - there's a webhook that will inject
a secret token + some environment variables if a Pod has a Service
Account set up.

paasta-tools will be in charge of creating the Service Account that
we'll use if it doesn't exist (as well as setting up the annotations on
that Service Account).