Catch all exceptions in k8s submit_command() (#926)
It's entirely possible that creating a task_processing task (and/or
submitting one) can result in an exception. At the moment, this results
in the affected ActionRun getting stuck in the Starting state - but this
is a lie and means that the normal monitoring/alerting on failed runs
does not kick in.