Add warnings about tronctl disable not being permanent
As we learned in a recent incident, if there are any issues with the
DynamoDB data, it's possible for us to end up re-enabling jobs. We've
previously warned folks in the past (in Slack/IRC/etc) that to
permanently disable jobs it's recommended that they do so in soaconfigs,
but this is easy to forget and not everyone will reach out before
disabling a job that they don't expect to re-enable any time soon.