TRON-2208: Add toggle in tron config to disable retries on LOST k8s jobs (#988)
Given that "LOST" means Tron has lost track of a pod it already thought it had started for a job, attempting to retry/start a replacement can be dangerous for non-idempotent jobs. In the current state, these will consume retries, but with some of our EKS migration methods, LOST tasks are more likely. Therefore, we should have a way to temporarily pause retries on these.

## Related Issues
- TRON-2208: Add toggle in Tron config to disable retries on LOST k8s jobs