Start running mypy over all of Tron
We previously only mypy'd a small subset of (mostly) fully typed files,
but we should probably start expanding the checked files - as well as
incrementally tightening up out mypy settings.

This PR enables type-checking all of Tron with the existing settings as
well as includes the disabled checks that I'll be attmepting to work
through during our unhack this week.