Default sliders to show 50 items (#1067)
Defaulting to 10 doesn't actually save any time since we've fetched all the data anyway.

Let's instead default to 50 since that should should be the most common run_limit

This will make pages longer, but it will be less confusing to folks when they get alerts for failed/unknown/etc jobs that are past the visible runs with the previous slider configuration.